Feat: 인기상품 조회 통합테스트 구현

+++ src/main/java/kr/hhplus/be/ecommerce/domain/product/ProductOrderRank.java
@@ -8,5 +8,5 @@ import lombok.Getter;
 public class ProductOrderRank {
 	private Integer productId;
 	private String productName;
-	private Long totalQuantity;
+	private Integer totalQuantity;
 }
